Got the push request working clientside and imput a lot of CSS. Will try to move the push request serverside.

+++ models/Product.js
@@ -1,35 +1,34 @@
-module.exports = function(connection, Sequelize) {
+module.exports = function (connection, Sequelize) {
 	const Product = connection.define('Product', {
-	 productName: {
-		type: Sequelize.STRING,
-		allowNull: false,
-		validate: {
-		  notEmpty: true
-		}
-	  },
-	  departmentName: {
-		type: Sequelize.TEXT,
-		allowNull: false,
-		validate: {
-		  notEmpty: true
-		}
-	  },
-	  price: {
-		type: Sequelize.INTEGER,
-		allowNull: false,
-		validate: {
-		  notEmpty: true
-		}
-	  },
-	  stock: {
-		type: Sequelize.INTEGER,
-		allowNull: false,
-		validate: {
-		  notEmpty: true
-		}
+		productName: {
+			type: Sequelize.STRING,
+			allowNull: false,
+			validate: {
+				notEmpty: true
+			}
+		},
+		departmentName: {
+			type: Sequelize.TEXT,
+			allowNull: false,
+			validate: {
+				notEmpty: true
+			}
+		},
+		price: {
+			type: Sequelize.FLOAT,
+			allowNull: false,
+			validate: {
+				notEmpty: true
+			}
+		},
+		stock: {
+			type: Sequelize.INTEGER,
+			allowNull: false,
+			validate: {
+				notEmpty: true
+			}
 		}
 	});
-  
+
 	return Product;
-  };
-  
+};
